fix: Rename Layout component entrypoint file
